+++ parser_library/src/expressions/conditional_assembly/terms/ca_string.cpp
@@ -14,9 +14,12 @@
 
 #include "ca_string.h"
 
+#include <optional>
+
 #include "expressions/conditional_assembly/ca_expr_visitor.h"
 #include "expressions/conditional_assembly/ca_operator_binary.h"
 #include "expressions/evaluation_context.h"
+#include "utils/unicode_text.h"
 
 namespace hlasm_plugin::parser_library::expressions {
 
@@ -72,11 +75,20 @@ bool ca_string::is_character_expression(character_expression_purpose) const { re
 
 void ca_string::apply(ca_expr_visitor& visitor) const { visitor.visit(*this); }
 
+namespace {
+bool string_too_long(std::string_view s, uint64_t dupl)
+{
+    return s.size() * dupl > utils::max_utf8_sequence_length * ca_string::MAX_STR_SIZE
+        || s.size() * dupl > ca_string::MAX_STR_SIZE
+        && utils::length_utf32_no_validation(s) * dupl > ca_string::MAX_STR_SIZE;
+}
+} // namespace
+
 context::SET_t ca_string::evaluate(const evaluation_context& eval_ctx) const
 {
     context::C_t str = semantics::concatenation_point::evaluate(value, eval_ctx);
 
-    if (str.size() > MAX_STR_SIZE)
+    if (string_too_long(str, 1))
     {
         eval_ctx.diags.add_diagnostic(diagnostic_op::error_CE011(expr_range));
         return context::object_traits<context::C_t>::default_v();
@@ -85,30 +97,34 @@ context::SET_t ca_string::evaluate(const evaluation_context& eval_ctx) const
     if (substring.start)
     {
         auto start = substring.start->evaluate(eval_ctx).access_a();
-        auto count =
-            substring.count ? substring.count->evaluate(eval_ctx).access_a() : (context::A_t)str.size() - start + 1;
+        std::optional<context::A_t> count;
+        if (substring.count)
+            count = substring.count->evaluate(eval_ctx).access_a();
 
         if (count == 0)
         {
             // when zero-length substring is requested, validation of the first parameter seems suppressed
             str = "";
         }
-        else if (start < 0 || count < 0 || (start == 0 && count > 0))
+        else if (start <= 0 || count && count < 0)
         {
             eval_ctx.diags.add_diagnostic(diagnostic_op::error_CE008(substring.substring_range));
             return context::object_traits<context::C_t>::default_v();
         }
-        else if (start > (context::A_t)str.size())
-        {
-            eval_ctx.diags.add_diagnostic(diagnostic_op::error_CE009(substring.start->expr_range));
-            return context::object_traits<context::C_t>::default_v();
-        }
-        /* TODO uncomment when assembler options will be implemented
-        if (start + count - 1 > (int)str.size())
-            eval_ctx.diags.add_diagnostic(diagnostic_op::error_CW001(substring.count->expr_range));
-        */
         else
-            str = str.substr(start - 1, count);
+        {
+            auto substr = utils::utf8_substr<false>(str, start - 1, (size_t)count.value_or(-1));
+            if (!substr.offset_valid || count && substr.str.empty())
+            {
+                eval_ctx.diags.add_diagnostic(diagnostic_op::error_CE009(substring.start->expr_range));
+                return context::object_traits<context::C_t>::default_v();
+            }
+            /* TODO uncomment when assembler options will be implemented
+            if (substr.char_count != count)
+                eval_ctx.diags.add_diagnostic(diagnostic_op::error_CW001(substring.count->expr_range));
+            */
+            str = std::string(substr.str);
+        }
     }
 
     return duplicate(duplication_factor, std::move(str), expr_range, eval_ctx);
@@ -127,7 +143,7 @@ std::string ca_string::duplicate(
             return "";
         }
 
-        if (value.size() * dupl > MAX_STR_SIZE)
+        if (string_too_long(value, (uint64_t)dupl))
         {
             eval_ctx.diags.add_diagnostic(diagnostic_op::error_CE011(expr_range));
             return "";