+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -25,6 +25,7 @@
 - Self-defining terms are incorrectly parsed in machine expressions
 - Incorrect ORG instruction behavior when operands contain EQUs
 - Library diagnostics point to a respective configuration file
+- Invalid substring may be generated when conditional assembly string contains multibyte UTF-8 characters
 
 ## [1.4.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.3.0...1.4.0) (2022-08-30)